refactor(mobile): collapse Wispr cluster (4 → 2 files)
The `wispr-*` cluster split a single feature across four files in
`apps/mobile/src/lib/`. Only `-automation-native.ts` was a real boundary
(Android accessibility-service JNI seam); the other three were pure
TypeScript with one prod caller (`detail.tsx`) and a type-only import
from `TextEntryModal.tsx`.

Merge `-automation-state.ts` (state machine reducer) + `-tap-timeout.ts`
(timeout utilities) + `-text-editor-flow.ts` (availability + control +
auto-close coordination) into a single `wispr-automation.ts`. Keep
`wispr-automation-native.ts` as the platform seam.

Test files mirror the source collapse: three `wispr-*.test.ts` files
merged into `wispr-automation.test.ts` (20 tests, same coverage).

External API preserved: `detail.tsx` imports collapse 4 → 2;
`TextEntryModal.tsx` `TextEntryWisprControl` type import updated.

typecheck clean, 381/381 integration tests pass.
