+++ apps/mobile/test/integration/wispr-automation.test.ts
@@ -1,13 +1,145 @@
 import assert from 'node:assert/strict';
 import test from 'node:test';
-
 import {
 	canStartWisprTextEntryAutomation,
+	reduceWisprAutomationState,
+	resolveTextEntryWisprControl,
 	resolveWisprAutoCloseOnTextEntryClose,
 	resolveWisprPendingAutoCloseRequests,
-	resolveTextEntryWisprControl,
 	resolveWisprTextEditorAvailability,
-} from '../../src/lib/wispr-text-editor-flow';
+	tapWisprControlWithTimeout,
+	WisprTapTimeoutError,
+	withTimeout,
+	type WisprAutomationState,
+} from '../../src/lib/wispr-automation';
+
+void test('first press opens and focuses text entry before starting Wispr', () => {
+	const initial: WisprAutomationState = { phase: 'idle' };
+
+	const opening = reduceWisprAutomationState(initial, { type: 'press' });
+	assert.deepEqual(opening, { phase: 'openingTextEntry' });
+
+	const waiting = reduceWisprAutomationState(opening, {
+		type: 'textEntryFocused',
+		textBeforeStart: '',
+	});
+	assert.deepEqual(waiting, {
+		phase: 'waitingForBubble',
+		textBeforeStart: '',
+	});
+
+	const recording = reduceWisprAutomationState(waiting, {
+		type: 'wisprTapSucceeded',
+	});
+	assert.deepEqual(recording, {
+		phase: 'recording',
+		textBeforeStart: '',
+	});
+});
+
+void test('recording ignores repeated presses and text change returns to idle', () => {
+	const recording: WisprAutomationState = {
+		phase: 'recording',
+		textBeforeStart: 'before',
+	};
+
+	const stillRecording = reduceWisprAutomationState(recording, {
+		type: 'press',
+	});
+	assert.deepEqual(stillRecording, {
+		phase: 'recording',
+		textBeforeStart: 'before',
+	});
+
+	const done = reduceWisprAutomationState(stillRecording, {
+		type: 'textChanged',
+		value: 'before dictated text',
+	});
+	assert.deepEqual(done, { phase: 'idle' });
+});
+
+void test('timeout records a retryable failure', () => {
+	const waiting: WisprAutomationState = {
+		phase: 'waitingForBubble',
+		textBeforeStart: '',
+	};
+
+	const failed = reduceWisprAutomationState(waiting, {
+		type: 'failed',
+		reason: 'bubble-not-found',
+		message: 'Wispr bubble not found',
+	});
+
+	assert.deepEqual(failed, {
+		phase: 'failed',
+		reason: 'bubble-not-found',
+		message: 'Wispr bubble not found',
+	});
+});
+
+const nextTick = () =>
+	new Promise<void>((resolve) => {
+		setImmediate(resolve);
+	});
+
+void test('Wispr tap timeout reports late native success', async () => {
+	let resolveTap: (value: string) => void = () => {};
+	let lateSuccessCount = 0;
+	const tapPromise = new Promise<string>((resolve) => {
+		resolveTap = resolve;
+	});
+
+	await assert.rejects(
+		tapWisprControlWithTimeout({
+			tapWisprControl: () => tapPromise,
+			timeoutMs: 1,
+			onLateSuccess: () => {
+				lateSuccessCount += 1;
+			},
+		}),
+		WisprTapTimeoutError,
+	);
+
+	resolveTap('ok');
+	await nextTick();
+
+	assert.equal(lateSuccessCount, 1);
+});
+
+void test('Wispr tap timeout reports late native failure', async () => {
+	let rejectTap: (error: Error) => void = () => {};
+	let lateFailureCount = 0;
+	const tapPromise = new Promise<string>((_, reject) => {
+		rejectTap = reject;
+	});
+
+	await assert.rejects(
+		tapWisprControlWithTimeout({
+			tapWisprControl: () => tapPromise,
+			timeoutMs: 1,
+			onLateFailure: () => {
+				lateFailureCount += 1;
+			},
+		}),
+		WisprTapTimeoutError,
+	);
+
+	rejectTap(new Error('native tap failed'));
+	await nextTick();
+
+	assert.equal(lateFailureCount, 1);
+});
+
+void test('withTimeout resolves before the deadline', async () => {
+	assert.equal(await withTimeout(Promise.resolve('ok'), 10), 'ok');
+});
+
+void test('withTimeout rejects when the wrapped promise hangs', async () => {
+	await assert.rejects(
+		withTimeout(new Promise<string>(() => {}), 1),
+		WisprTapTimeoutError,
+	);
+});
 
 void test('disabled Wispr service keeps text editor usable without opening settings', () => {
 	const result = resolveWisprTextEditorAvailability({
